A Simple Algorithmic Trading and Backtesting Model.
This gives the user a few options for backtesting models. You can
select the following:

SMA: 5, 15, or 30 days
Above/Below: Choose to buy when the day is newly above or below the SMA.
Days to Hold: How many days to hold the asset.

The data provided is 252 days of GSPC daily data, but can be changed as
needed.